+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * :warning: remove deprecated `ReverseSlice` function in `basicalter`
 * :warning: remove deprecated `EqualSlice` function in `basiccheck`
 * :warning: remove deprecated `InSlice` function in `basiccheck`
+* :warning: remove deprecated `OneInSliceWith` function in `basiccheck`
 
 ## v0.12.0